+++ .gitignore
@@ -1,0 +1,5 @@
+.vs/
+CMakeSettings.json
+bin/
+lib/
+linux/
