+++ bitpolymul2/bc.h
