+++ bitpolymul2/bc_to_gen_code.h
