+++ bitpolymul2/bc_to_lch_gen_code.c
