+++ bitpolymul2/bc_to_mono_gen_code.c
