+++ bitpolymul2/bitpolymul.c
