+++ bitpolymul2/byte_inline_func.h
@@ -23,7 +23,7 @@ along with BitPolyMul.  If not, see <http://www.gnu.org/licenses/>.
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdint.h>
-#include "defines.h"
+#include "bitpolymul2/defines.h"
 
 
 static inline
