+++ bitpolymul2/defines.h
