+++ bitpolymul2/encode.c
