+++ bitpolymul2/encode.h
