+++ bitpolymul2/gf2128_cantor_iso.h
