+++ bitpolymul2/gf264_cantor_iso.c
