+++ bitpolymul2/ska.h
