+++ bitpolymul2/trunc_btfy_tab.c
