+++ bitpolymul2/trunc_btfy_tab_64.h
