+++ bitpolymul2tests/CMakeLists.txt
@@ -1,0 +1,5 @@
+
+
+add_executable(bitpolymul-test "${CMAKE_CURRENT_SOURCE_DIR}/bitpolymul-test.cpp")
+
+target_link_libraries(bitpolymul-test bitpolymul2)
