+++ bitpolymul2tests/bitpolymul-test.cpp
