+++ run_test.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 
-./bitpolymul-test 15
-./bitpolymul-test 16
-./bitpolymul-test 17
-./bitpolymul-test 18
-./bitpolymul-test 19
-./bitpolymul-test 20
-./bitpolymul-test 21
-./bitpolymul-test 22
-./bitpolymul-test 23
+./bin/bitpolymul-test 15
+./bin/bitpolymul-test 16
+./bin/bitpolymul-test 17
+./bin/bitpolymul-test 18
+./bin/bitpolymul-test 19
+./bin/bitpolymul-test 20
+./bin/bitpolymul-test 21
+./bin/bitpolymul-test 22
+./bin/bitpolymul-test 23
